Prepare for the return of FTP target support.
The backup target is now specified by the target.uri config variable,
which is a Duplicity URI.

If target.mount is set then it will be mounted during the backup run. It
should be the name of a device or directory configured in /etc/fstab
such that 'mount ${target.mount}' will succeed.