Don't redirect stdout/stderr of launched commands
Since the output is no longer buffered, when run from a terminal command
output is displayed in real-time. Flushing sys.stdout and sys.stderr
before commands are run should make sure that output is ordered
correctly.

stdout is still captured if the caller asks for it; this allows the FTP
mounter to make use of it when it invokes lftp(1) to calculate the used
space on an FTP server.