Add a separate config option for specifying the backup target by device name
The mount class now prefers to deal with targets specified by mount
point; if it is only given a device it will use getfsspec(3) to
determine where the device should be mounted.

This is necessary becase on some hosts (e.g., when backing up to an NFS
mount), we specify the target in terms of the mount point, but on others
(e.g., when backing up to disk), we specify the target in terms of the
device to be mounted.